Improved the git2gdrive script
It will now check if you have the drive utility in your path. If you do not
it will refuse to run. This was an issue when setting it up as a cronjob,
because `drive` was not in my path originally. I also improved the
documentation within the script. I also removed the directory manipulations
from the script, because that is not required anymore because I discovered
`git archive`, and the script has been using that for a while now.