Worked on the plusMinus Graph script
The insert and delete calculation script is done to my satisfaction.
It now includes a net lines calculation and respect to time for the drawing
the graph. It will probably be modified in the future, but I don't really
care right now. There is an issue with the creation of the graph labels,
but that is not really important.