Added an optional repository sync setting to the ghist script
I added an option repository sync flag to the ghist script. This does not
change existing behavior of the script. It is entirely up to you if you
want to enable it. It mearly uses `git fetch` to grab the updates from the
remote if you have one. It should be noted that this assumes that the
remote is named `origin` and the main branch is named `master`, there are
plans to make this configurable by the user.